addressed feedback. LMK if it looks good
