+++ frontend/src/pages/Account.tsx
@@ -2,200 +2,135 @@ import SESDCHeader from '../components/SESDCHeader';
 import SESDCFooter from '../components/SESDCFooter';
 
 import { useState } from 'preact/hooks';
-import { deleteAccount } from '../utils/deleteAccount';
-import { auth } from '../utils/firebase/firebase-init';
-import { updatePassword, EmailAuthProvider, reauthenticateWithCredential } from 'firebase/auth';
-
 import '../css/account.css';
 
+import { deleteAccount } from "../utils/deleteAccount";
+import { updateUserPassword, logoutUser } from "../utils/user";
+
 export default function Account() {
-  const [showUpdatePassword, setShowUpdatePassword] = useState(false);
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [message, setMessage] = useState('');
+    const [showUpdatePassword, setShowUpdatePassword] = useState(false);
+    const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
+    const [message, setMessage] = useState('');
 
-  function toggleUpdatePassword() {
-    setShowUpdatePassword((prev) => !prev);
-    setShowDeleteConfirm(false);
-    setMessage('');
-  }
-
-  function toggleDelete() {
-    setShowDeleteConfirm((prev) => !prev);
-    setShowUpdatePassword(false);
-    setMessage('');
-  }
-
-  // update password 
-  async function handleConfirmPasswordUpdate() {
-    const newPasswordInput = document.getElementById('newPassword') as HTMLInputElement;
-    const newPassword = newPasswordInput.value;
-
-    if (!newPassword) {
-      setMessage("Please enter a new password.");
-      return;
+    function toggleUpdatePassword() {
+        setShowUpdatePassword(prev => !prev);
+        setShowDeleteConfirm(false);
+        setMessage('');
     }
 
-    try {
-      const user = auth.currentUser;
-      if (!user) {
-        setMessage("No authenticated user.");
-        return;
-      }
-
-      // Optionally reauth with old password (Firebase now requires reauth for this)
-      // For now, just update:
-      await updatePassword(user, newPassword);
-
-      setMessage("Password updated successfully.");
-      setShowUpdatePassword(false);
-    } catch (error: any) {
-      setMessage("Error updating password: " + error.message);
-    }
-  }
-
-  function handleCancelPasswordUpdate() {
-    setShowUpdatePassword(false);
-  }
-
-  //delete account
-  async function handleConfirmDelete() {
-    const passwordInput = document.getElementById('deletePassword') as HTMLInputElement;
-    const password = passwordInput.value;
-
-    if (!password) {
-      setMessage("Please enter your password.");
-      return;
+    function toggleDelete() {
+        setShowDeleteConfirm(prev => !prev);
+        setShowUpdatePassword(false);
+        setMessage('');
     }
 
-    const result = await deleteAccount(password);
-    setMessage(result);
+    async function handleConfirmPasswordUpdate() {
+        const passwordInput = document.getElementById('newPassword') as HTMLInputElement;
+        const newPassword = passwordInput.value;
 
-    if (result === "Account deleted successfully.") {
-      setTimeout(() => {
-        window.location.href = '/login';
-      }, 1500);
+        const response = await updateUserPassword(newPassword);
+        setMessage(response);
+
+        if (response === "Password updated successfully.") {
+            setShowUpdatePassword(false);
+            passwordInput.value = "";
+        }
     }
-  }
 
-  function handleCancelDelete() {
-    setShowDeleteConfirm(false);
-  }
+    function handleCancelPasswordUpdate() {
+        setShowUpdatePassword(false);
+    }
 
-  // logout for “Go to Login”
-  async function handleLogoutAndGoToLogin() {
-    await auth.signOut();
-    localStorage.removeItem("loggedInUserId");
-    window.location.href = "/login";
-  }
+    async function handleConfirmDelete() {
+        const passwordInput = document.getElementById('deletePassword') as HTMLInputElement;
+        const password = passwordInput.value;
 
-  return (
-    <>
-      <SESDCHeader />
+        const result = await deleteAccount(password);
+        setMessage(result);
 
-      <div class="main-content">
-        <div class="wrapper">
-          <div id="accountManagement" class="form-container visible">
-            <h1>Account Management</h1>
+        if (result === "Account deleted successfully.") {
+            setTimeout(async () => {
+                await logoutUser();
+                window.location.href = "/login";
+            }, 2000);
+        }
+    }
 
-            {message && (
-              <div class="messageDiv" style="display: block;">
-                {message}
-              </div>
-            )}
+    function handleCancelDelete() {
+        setShowDeleteConfirm(false);
+    }
 
-            <div class="button-group">
-              {/* UPDATE PASSWORD */}
-              <button
-                class="btn secondary-btn"
-                onClick={toggleUpdatePassword}
-                type="button"
-              >
-                {showUpdatePassword ? 'Hide Password Form' : 'Update Password'}
-              </button>
+    async function handleGoToLogin() {
+        await logoutUser();
+        window.location.href = "/login";
+    }
 
-              {showUpdatePassword && (
-                <div id="updatePasswordSection">
-                  <div class="input-box">
-                    <input
-                      id="newPassword"
-                      type="password"
-                      placeholder="Enter new password"
-                      required
-                    />
-                  </div>
-                  <div class="confirmation-buttons">
-                    <button
-                      class="btn secondary-btn"
-                      type="button"
-                      onClick={handleConfirmPasswordUpdate}
-                    >
-                      Confirm Update
-                    </button>
-                    <button
-                      class="btn secondary-btn"
-                      type="button"
-                      onClick={handleCancelPasswordUpdate}
-                    >
-                      Cancel
-                    </button>
-                  </div>
+    return (
+        <>
+            <SESDCHeader />
+
+            <div class="main-content">
+                <div class="wrapper">
+                    <div id="accountManagement" class="form-container visible">
+                        <h1>Account Management</h1>
+
+                        {message && (
+                            <div class="messageDiv" style="display: block;">
+                                {message}
+                            </div>
+                        )}
+
+                        <div class="button-group">
+                            <button class="btn secondary-btn" type="button" onClick={toggleUpdatePassword}>
+                                {showUpdatePassword ? 'Hide Password Form' : 'Update Password'}
+                            </button>
+
+                            {showUpdatePassword && (
+                                <div id="updatePasswordSection">
+                                    <div class="input-box">
+                                        <input id="newPassword" type="password" placeholder="Enter new password" required />
+                                    </div>
+                                    <div class="confirmation-buttons">
+                                        <button class="btn secondary-btn" type="button" onClick={handleConfirmPasswordUpdate}>
+                                            Confirm Update
+                                        </button>
+                                        <button class="btn secondary-btn" type="button" onClick={handleCancelPasswordUpdate}>
+                                            Cancel
+                                        </button>
+                                    </div>
+                                </div>
+                            )}
+
+                            <button class="btn danger-btn" type="button" onClick={toggleDelete}>
+                                {showDeleteConfirm ? 'Hide Delete Form' : 'Delete Account'}
+                            </button>
+
+                            <button class="btn primary-btn" type="button" onClick={handleGoToLogin}>
+                                Go to Login
+                            </button>
+                        </div>
+
+                        {showDeleteConfirm && (
+                            <div id="deleteConfirmation">
+                                <p>Are you sure you want to delete your account? This action cannot be undone.</p>
+                                <div class="input-box">
+                                    <input id="deletePassword" type="password" placeholder="Enter your password to confirm" required />
+                                </div>
+                                <div class="confirmation-buttons">
+                                    <button class="btn danger-btn" type="button" onClick={handleConfirmDelete}>
+                                        Confirm Delete
+                                    </button>
+                                    <button class="btn secondary-btn" type="button" onClick={handleCancelDelete}>
+                                        Cancel
+                                    </button>
+                                </div>
+                            </div>
+                        )}
+                    </div>
                 </div>
-              )}
-
-              {/* DELETE ACCOUNT */}
-              <button
-                class="btn danger-btn"
-                type="button"
-                onClick={toggleDelete}
-              >
-                {showDeleteConfirm ? 'Hide Delete Form' : 'Delete Account'}
-              </button>
-
-              <button
-                class="btn primary-btn"
-                type="button"
-                onClick={handleLogoutAndGoToLogin}
-              >
-                Go to Login
-              </button>
             </div>
 
-            {showDeleteConfirm && (
-              <div id="deleteConfirmation">
-                <p>
-                  Are you sure you want to delete your account? This action cannot be undone.
-                </p>
-                <div class="input-box">
-                  <input
-                    id="deletePassword"
-                    type="password"
-                    placeholder="Enter your password to confirm"
-                    required
-                  />
-                </div>
-                <div class="confirmation-buttons">
-                  <button
-                    class="btn danger-btn"
-                    type="button"
-                    onClick={handleConfirmDelete}
-                  >
-                    Confirm Delete
-                  </button>
-                  <button
-                    class="btn secondary-btn"
-                    type="button"
-                    onClick={handleCancelDelete}
-                  >
-                    Cancel
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-
-      <SESDCFooter />
-    </>
-  );
+            <SESDCFooter />
+        </>
+    );
 }
